Make 'sourcemaps: existing' the default to match user expectations
Reasoning: If the user already has a source map for a CSS file (either static or generated),
there is a high probability that they still want to have a functioning source map
after Autoprefixer has been applied.

Backwards compatibility: No files will be generated that weren't there in earlier versions;
source map files that were present but unusable because they didn't match
the transformed CSS will be fixed after updating the plugin.

+++ lib/jekyll/autoprefixer/autoprefixer.rb
@@ -25,9 +25,9 @@ module Jekyll
           case options['sourcemaps']
           when 'never', false then false
           when 'always', true then true
+          when 'existing', nil then :transform_existing_only # Default value
           when 'production' then Jekyll.env == 'production'
           when 'development' then Jekyll.env == 'development'
-          when nil then false # Default value
           else
             Jekyll.logger.warn("Ignoring unknown value '#{options['sourcemaps']}' for autoprefixer.sourcemaps. Disabling source map generation instead.")
             false
@@ -69,17 +69,20 @@ module Jekyll
 
             map_file = File.open(map_path, 'r+')
             map_options = { 'prev' => map_file.read, 'inline' => false }
+            map_expected = true
 
-          elsif write_sourcemaps
+          elsif write_sourcemaps && write_sourcemaps != :transform_existing_only
             Jekyll.logger.debug 'Autoprefixer:', "Creating new map: #{map_path}"
 
             map_file = File.open(map_path, 'w+')
             map_options = { 'inline' => false }
+            map_expected = true
 
           else
             # No sourcemaps should be written
             map_file = nil
             map_options = nil
+            map_expected = false
           end
 
           file_options = options
@@ -95,7 +98,7 @@ module Jekyll
             map_file.truncate(0)
             map_file.rewind
             map_file.write(result.map)
-          elsif write_sourcemaps
+          elsif map_expected
             Jekyll.logger.error 'Autoprefixer Error:', "Failed to create sourcemap for #{filename} found at path #{path}"
           end
         ensure
